[net] placed hardcodes `from/to` near by in the code
